implemented changes in audio and surface

+++ src/SDL/___CLEANED/stdinc/SDL_stdinc.cs
@@ -12,9 +12,10 @@ unsafe partial class SDL
 		return (uint)(a | (b << 8) | (c << 16) | (d << 24));
 	}
 
+	// TODO: consider if adding malloc is a good idea.
 	[LibraryImport(LibraryName, EntryPoint = "SDL_malloc")]
 	[UnmanagedCallConv(CallConvs = [typeof(CallConvCdecl)])]
-	public static partial void* Malloc(ulong size); // adding this is probably a bad idea
+	public static partial void* Malloc(nuint size);
 
 	[LibraryImport(LibraryName, EntryPoint = "SDL_free")]
 	[UnmanagedCallConv(CallConvs = [typeof(CallConvCdecl)])]
